Initial forklift commit of os_resource_classes
A start of a library for keeping standardized resource class information
for OpenStack. This includes a "version history" that comes from the
original nova.objects.fields.ResourceClassField Enum field.

The only resource class codes and names in use at the time of this
forklift are:

- VCPU
- MEMORY_MB
- DISK_GB
- VGPU

These other resource classes are not used in Nova, and this commit
deprecates those resource class names and their associated indexes/codes
from the original nova ResourceClassField enum type:

- PCI_DEVICE
- SRIOV_NET_VF
- NUMA_SOCKET
- NUMA_CORE
- NUMA_THREAD
- NUMA_MEMORY_MB
- IPV4_ADDRESS
- VGPU_DISPLAY_HEAD

The intention is to add new resource classes for those resources that we
are actively going to be tracking with the placement service as needed,
and use a namespace/prefix to indicate categories of resources (e.g.
we'd like to use a NET_ prefix for the SR-IOV VF network interface and
IPv4 address resource classes).